refactor: new util file for llm.py, rename test file
