+++ .env.example
@@ -2,6 +2,6 @@
 AI_SERVER_URL="http://localhost:8000"
 
 # Discord Bot
-DISCORD_BOT_TOKEN="EXAMPLETokenTTTTT.Grz0fK.IUi1IONPcLHZsnyjROTt8lR2fuYCeK4grzRoSQ"
+DISCORD_BOT_TOKEN="EXAMPLETokennnnnn.Grz0fK.IUi1IONPcLHZsnyjROTt8lR2fuYCeK4grzRoSQ"
 # THIS SHOULD BE ASSIGNED TO THE BOT'S `scopes` WHEN DEPLOYED
 DISCORD_GUILD_ID="99999999999999999999999"
