+++ src/discord_bot/llm.py
@@ -7,11 +7,7 @@ by modifying the `base_url`.
 import os
 import openai
 import re
-from utils import pdf
-
-MAX_CHARACTERS = 2000
-QUESTION_CUT_OFF_LENGTH = 150
-RESERVED_SPACE = 50  # for other additional strings. E.g. number `(1/4)`, `Q: `, `A: `, etc.
+from utils import llm_response, pdf
 
 
 async def answer_question(model: str, question: str, server_url: str, attach_question_to_message: bool = True) -> list[str]:
@@ -25,30 +21,30 @@ async def answer_question(model: str, question: str, server_url: str, attach_que
       messages=[{"role": "user", "content": question}],
     )
     content = response.choices[0].message.content or "No response from the model. Please try again"
-    messages = split(content)
-    messages = add_number(messages)
+    messages = llm_response.split(content)
+    messages = llm_response.add_number(messages)
     if attach_question_to_message:
-      messages = add_question(messages, question)
+      messages = llm_response.add_question(messages, question)
 
     return messages
 
   except Exception as e:
-    return split(f"Error in calling the LLM: {e}")
+    return llm_response.split(f"Error in calling the LLM: {e}")
 
 
 async def review_resume(model: str, url: str, server_url: str) -> list[str]:
   # validate url structure, must have leading "http[s]?" and a domain name (e.g. "example.com"=`\w+\.\w+`)
   if not re.search(r"http[s]?://\w+\.\w+", url):
-    return split("Invalid URL. Please provide a valid URL of your resume.")
+    return llm_response.split("Invalid URL. Please provide a valid URL of your resume.")
 
   output_path = ""
   try:
     output_path = pdf.download(url)
     if not pdf.validate_pdf_format(output_path):
-      return split("Error: Invalid PDF format. Please provide a valid PDF file.")
+      return llm_response.split("Error: Invalid PDF format. Please provide a valid PDF file.")
     text = pdf.parse_to_text(output_path)
   except Exception as e:
-    return split(f"Error in processing PDF: {e}")
+    return llm_response.split(f"Error in processing PDF: {e}")
   finally:
     if os.path.exists(output_path):
       os.remove(output_path)
@@ -61,48 +57,4 @@ async def review_resume(model: str, url: str, server_url: str) -> list[str]:
   )
 
   messages = await answer_question(model, question, server_url, attach_question_to_message=False)
-  return messages
-
-
-def split(answer: str) -> list[str]:
-  """
-  Split the answer into a list of smaller strings so that
-  each element is less than MAX_CHARACTERS characters.
-  Full sentences are preserved.
-  """
-  limit = MAX_CHARACTERS - RESERVED_SPACE - QUESTION_CUT_OFF_LENGTH
-  messages = []
-  answer = answer.strip()
-
-  while len(answer) > limit:
-    last_period = answer[:limit].rfind(".")
-    if last_period == -1:
-      last_period = answer[:limit].rfind(" ")
-    messages.append(answer[: last_period + 1])
-    answer = answer[last_period + 1 :]
-
-  messages.append(answer)
-
-  return messages
-
-
-def add_question(messages: list[str], questions: str) -> list[str]:
-  """
-  Add the asked question to the beginning of each message.
-  """
-  return [(f"Q: {questions[:QUESTION_CUT_OFF_LENGTH]}\n" + f"A: {message}") for message in messages]
-
-
-def add_number(messages: list[str]) -> list[str]:
-  """
-  Add the number to the beginning of each message. E.g. `(1/4)`
-  Do nothing if the length of `messages` is 1.
-  """
-  if len(messages) == 1:
-    return messages
-
-  for i, message in enumerate(messages):
-    message = message.strip()
-    messages[i] = f"({i+1}/{len(messages)}) {message}"
-
   return messages
