+++ src/utils/llm_response.py
@@ -1,0 +1,47 @@
+DISCORD_MESSAGE_MAX_CHARACTERS = 2000
+QUESTION_CUT_OFF_LENGTH = 150
+RESERVED_LENGTH = 50  # for other additional strings. E.g. number `(1/4)`, `Q: `, `A: `, etc.
+
+
+def split(answer: str) -> list[str]:
+  """
+  Split the answer into a list of smaller strings so that
+  each element is less than discord's message length limit.
+  Full sentences are preserved where possible.
+  """
+  limit = DISCORD_MESSAGE_MAX_CHARACTERS - RESERVED_LENGTH - QUESTION_CUT_OFF_LENGTH
+  messages = []
+  answer = answer.strip()
+
+  while len(answer) > limit:
+    last_period = answer[:limit].rfind(".")
+    if last_period == -1:
+      last_period = answer[:limit].rfind(" ")
+    messages.append(answer[: last_period + 1])
+    answer = answer[last_period + 1 :]
+
+  messages.append(answer)
+
+  return messages
+
+
+def add_question(messages: list[str], questions: str) -> list[str]:
+  """
+  Add the asked question to the beginning of each message.
+  """
+  return [(f"Q: {questions[:QUESTION_CUT_OFF_LENGTH]}\n" + f"A: {message}") for message in messages]
+
+
+def add_number(messages: list[str]) -> list[str]:
+  """
+  Add the number to the beginning of each message. E.g. `(1/4)`
+  Do nothing if the length of `messages` is 1.
+  """
+  if len(messages) == 1:
+    return messages
+
+  for i, message in enumerate(messages):
+    message = message.strip()
+    messages[i] = f"({i+1}/{len(messages)}) {message}"
+
+  return messages
